Cleaned up 64bit errors

+++ victorious/victorious/UI/VProfileViewController.m
@@ -138,7 +138,7 @@
     [actionSheet showFromBarButtonItem:sender animated:YES];
 }
 
--(void)actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(int)buttonIndex
+-(void)actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex
 {
     if (buttonIndex == 0)
     {
